Rename volumes to zfs pools
Ticket: #26111
