--- src/app/pages/storage/volumes/datasets/dataset-delete/index.ts
@@ -1,1 +1,0 @@
-export {DatasetDeleteComponent} from './dataset-delete.component';
