--- src/app/pages/storage/volumes/datasets/dataset-form/index.ts
@@ -1,1 +1,0 @@
-export { DatasetFormComponent } from './dataset-form.component';
