--- src/app/pages/storage/volumes/datasets/dataset-permissions/index.ts
@@ -1,1 +1,0 @@
-export {DatasetPermissionsComponent} from './dataset-permissions.component';
