--- src/app/pages/storage/volumes/import-list/index.ts
@@ -1,1 +1,0 @@
-export {VolumeImportListComponent} from './volume-import.component';
