--- src/app/pages/storage/volumes/manager/disk/index.ts
@@ -1,1 +1,0 @@
-export * from './disk.component';
