--- src/app/pages/storage/volumes/manager/index.ts
@@ -1,3 +1,0 @@
-export * from './manager.component';
-export * from './disk/disk.component';
-export * from './vdev/vdev.component';
