--- src/app/pages/storage/volumes/manager/manager.component.css
@@ -1,33 +1,0 @@
-.custom_cell_styling {
-	overflow: hidden; 
-	height: 20px; 
-	width: 100%; 
-	padding-left: 5px;
-}
-
-.wrapper {
-	display: flex;
-	flex-flow: row wrap;
-	-webkit-box-orient: horizontal;
-	-webkit-box-direction: normal;
-}
-
-.disks {
-	flex-basis: auto;
-	-webkit-box-flex: 1;
-	flex-grow: 1;
-	min-width: 350px; 
-	max-width: 400px;
-}
-.pool {
-	flex-basis: auto;
-	-webkit-box-flex: 1;
-	flex-grow: 1;
-	min-width: 400px;
-	max-width: 450px;
-}
-
-h4 {
-	text-align: center;
-}
-
