--- src/app/pages/storage/volumes/manager/manager.component.html
@@ -1,103 +1,0 @@
-<md-card class="m-0 ml-1 mr-1 p-0">
-  <div class="mat-toolbar mat-card-toolbar">
-    <div class="mat-card-title-text">ZFS Pool Manager</div>
-  </div>
-  <div class="md-content p-1">
-    <div class="m-3">
-      <div *ngIf="isNew;else extend">
-      <md-input-container>
-          <input mdInput placeholder="Name" [(ngModel)]="name" required>
-      </md-input-container>
-        <md-checkbox
-          [(ngModel)]="isEncrypted"
-          class="form-control"
-           placeholder="Encryption"
-           (change)="isEncryptedChecked()" >
-           Encryption
-        </md-checkbox>
-        <div>
-          <button *ngIf="suggestable_disks.length <= 8" [disabled]="dirty" md-raised-button class="btn btn-accent" (click)="suggestLayout()">Suggest Layout</button>
-        </div>
-      </div>
-      <ng-template #extend>
-        <md-input-container>
-          <input mdInput placeholder="Name" [value]="name" disabled>
-        </md-input-container>
-        <md-error *ngIf="isEncrypted">Extending an encrypted volume will reset the passphrase and recovery key!</md-error>
-      </ng-template>
-    </div>
-    <div class="wrapper">
-      <div class="disks">
-          <h4>Available Disks</h4>
-            <ngx-datatable
-              #table
-              class="material ml-1 mt-1 mb-1 mr-1 mat-box-shadow"
-              [rows]="disks"
-              [headerHeight]="50"
-              [rowHeight]="50"
-              [footerHeight]="50"
-              [columnMode]="'force'"
-              [limit]="10"
-              [selected] = "selected"
-              [selectionType]="'checkbox'"
-              (select)='onSelect($event)'>
-              <ngx-datatable-column
-                [width]="30"
-                [sortable]="false"
-                [canAutoResize]="false"
-                [draggable]="false"
-                [resizeable]="false"
-                [headerCheckboxable]="true"
-                [checkboxable]="true">
-              </ngx-datatable-column>
-              <ngx-datatable-column name="Disk" prop="devname"></ngx-datatable-column>
-              <ngx-datatable-column name="Capacity" prop="capacity"></ngx-datatable-column>
-            </ngx-datatable>
-            <md-input-container>
-                <input mdInput
-                  id="nameFilter"
-                  type='text'
-                  placeholder='Filter disks by name'
-                  (keyup)='updateFilter($event)'
-                />
-              </md-input-container>
-              <md-input-container>
-                <input mdInput
-                  id="capacityFilter"
-                  type='text'
-                  placeholder='Filter disks by capacity'
-                  (keyup)='updateFilter($event)'
-                />
-              </md-input-container>
-              <md-error *ngIf="re_has_errors">{{re_errors}}</md-error>
-      </div>
-      <div class="pool">
-        <div>
-          <h4>Data VDevs</h4>
-          <app-vdev *ngFor="let vdev of vdevs.data; let i = index" [group]="'data'" [manager]="this" [index]="i"></app-vdev>
-        </div>
-        <div *ngIf="vdevs.cache.length > 0">
-          <h4>Cache VDev</h4>
-          <app-vdev *ngFor="let vdev of vdevs.cache; let i = index" [group]="'cache'" [manager]="this" [index]="i"></app-vdev>
-        </div>
-        <div *ngIf="vdevs.log.length > 0">
-          <h4>Log VDev</h4>
-          <app-vdev *ngFor="let vdev of vdevs.log; let i = index" [group]="'log'" [manager]="this" [index]="i"></app-vdev>
-        </div>
-        <div *ngIf="vdevs.spare.length > 0">
-          <h4>Spare VDev</h4>
-          <app-vdev *ngFor="let vdev of vdevs.spare; let i = index" [group]="'spare'" [manager]="this" [index]="i"></app-vdev>
-        </div>
-      </div>
-    </div>
-  <div class="buttons">
-  <div *ngIf="error" [style.color]="'red'" [innerHTML]="error"></div>
-        <button md-raised-button class="btn btn-primary" (click)="doSubmit()" color="primary">Save</button>
-        <button md-raised-button class="btn  btn-block btn-raised btn-lg btn-primary" type="button" (click)="goBack()" color="accent">Cancel</button>
-        <button md-raised-button class="btn btn-accent" (click)="addVdev('data')">Add Data</button>
-        <button [disabled]="vdevs.cache.length > 0" md-raised-button class="btn btn-accent" (click)="addVdev('cache')">Add Cache</button>
-        <button [disabled]="vdevs.log.length > 0" md-raised-button class="btn btn-accent" (click)="addVdev('log')">Add Log</button>
-        <button [disabled]="vdevs.spare.length > 0" md-raised-button class="btn btn-accent" (click)="addVdev('spare')">Add Spare</button>
-    </div>
-  </div>
-</md-card>
