--- src/app/pages/storage/volumes/manager/vdev/index.ts
@@ -1,1 +1,0 @@
-export * from './vdev.component';
