--- src/app/pages/storage/volumes/manager/vdev/vdev.component.html
@@ -1,58 +1,0 @@
-<div class="vdev-wrapper">
-<div class="vdev-actions">
-    <button md-flat-button class="vdev-action-btn" (click)="addSelectedDisks()" [disabled]="!manager.selected.length"><md-icon>chevron_right</md-icon></button>
-    <button md-flat-button class="vdev-action-btn" (click)="removeSelectedDisks()" [disabled]="!selected.length"><md-icon>chevron_left</md-icon></button>
-</div>
-<div class="vdev-table">
-    <ngx-datatable
-        #table
-        class="material ml-1 mt-1 mb-1 mr-1 mat-box-shadow" 
-        [rows]="getDisks()" 
-        [headerHeight]="48" 
-        [rowHeight]="30"
-        [footerHeight]="48" 
-        [columnMode]="'flex'" 
-        [limit]="10"
-        [selected] = "selected"
-        [selectionType]="'checkbox'"
-        (select)='onSelect($event)'>
-        <ngx-datatable-column
-        	 [flexGrow]="1"
-             [width]="24"
-             [sortable]="false"
-             [canAutoResize]="false"
-             [draggable]="false"
-             [resizeable]="false"
-             [headerCheckboxable]="true"
-             [checkboxable]="true">
-        </ngx-datatable-column>
-        <ngx-datatable-column 
-           [flexGrow]="2"
-           name="Disk" 
-           prop="devname"></ngx-datatable-column>
-        <ngx-datatable-column 
-           	[flexGrow]="2"
-           	name="Capacity" 
-           	prop="capacity"></ngx-datatable-column>
-     </ngx-datatable>
-     <div class="vdev-wrapper">
-         <div class="vdev-type">
-         <div *ngIf="group === 'data'">
-            <md-select *ngIf="group === 'data'" [(ngModel)]="type" (change)="onTypeChange($event)" class="form-control">
-                <md-option value="stripe">Stripe</md-option>
-                <md-option value="mirror" *ngIf="getDisks().length >= 2">Mirror</md-option>
-                <md-option value="raidz" *ngIf="getDisks().length >= 3">Raid-z</md-option>
-                <md-option value="raidz2" *ngIf="getDisks().length >= 4">Raid-z2</md-option>
-                <md-option value="raidz3" *ngIf="getDisks().length >= 5">Raid-z3</md-option>
-            </md-select>
-            Estimated raw capacity: {{size}}
-        </div>
-        </div>
-        <div class="vdev-close">
-            <button md-flat-button class="vdev-action-btn" *ngIf="removable" (click)="remove()" [disabled]="(group === 'data' && manager.vdevs['data'].length < 2)">X</button>
-         </div>
-     </div>
-</div>
-</div>
-
-
