--- src/app/pages/storage/volumes/status/status.component.css
@@ -1,3 +1,0 @@
-.padding-16 {
-  padding: 16px;
-}
