--- src/app/pages/storage/volumes/volume-delete/index.ts
@@ -1,1 +1,0 @@
-export {VolumeDeleteComponent} from './volume-delete.component';
