--- src/app/pages/storage/volumes/volumes-list/index.ts
@@ -1,1 +1,0 @@
-export {VolumesListComponent} from './volumes-list.component';
