--- src/app/pages/storage/volumes/volumes-list/volumes-list.component.html
@@ -1,51 +1,0 @@
-<div id="volumes-list.component_html"  class="material mat-card mat-card-table">
-    <div class="mat-toolbar mat-card-toolbar">
-        <div class="mat-card-title-text">{{ title }}</div>
-        <app-entity-table-add-actions [entity]="this"></app-entity-table-add-actions>
-    </div>
-    
-    <div>
-        <br/>
-    </div>
-
-    <div *ngIf="zfsPoolRows.length > 0">
-    <md-accordion>
-        <md-expansion-panel *ngFor="let row of zfsPoolRows" [expanded]="expanded">
-          <md-expansion-panel-header>
-            <md-panel-title>
-                {{row.name}}
-            </md-panel-title>
-            <md-panel-description>
-                {{row.status}} ({{rowValue(row, 'avail')}} Free) 
-            </md-panel-description>
-          </md-expansion-panel-header>
-          <div>
-                <div style="margin-top: 8px; float: right; padding-bottom: 20px;">
-                    <app-entity-table-actions [icon_name]="'menu'" [entity]="this" [row]="row"></app-entity-table-actions>
-                </div>
-                <br/>
-          </div>
-
-          <div>
-              <br/>
-          </div>
-
-          <entity-table [conf]="row.volumesListTableConfig"></entity-table>
-
-
-        </md-expansion-panel>
-    </md-accordion>
-    </div>
-
-    <div *ngIf="zfsPoolRows.length < 1">
-        <md-card>
-            <md-card-content>
-                    No ZFS Pools found.
-            </md-card-content>
-        </md-card>
-        <br/>
-        <br/>
-    </div>
-
-
-</div>
