--- src/app/pages/storage/volumes/zvol/zvol-add/index.ts
@@ -1,1 +1,0 @@
-export {ZvolAddComponent} from './zvol-add.component';
