--- src/app/pages/storage/volumes/zvol/zvol-delete/index.ts
@@ -1,1 +1,0 @@
-export {ZvolDeleteComponent} from './zvol-delete.component';
