--- src/app/pages/storage/volumes/zvol/zvol-edit/index.ts
@@ -1,1 +1,0 @@
-export {ZvolEditComponent} from './zvol-edit.component';
